tutorials/qoa4ml/preprocessor/requirements.txt
flask==2.3.3
pillow==10.0.1
requests==2.31.0
pika==1.3.2
psutil
prometheus-client
paho-mqtt==1.6.1
pandas
qoa4ml==0.1.20
dnspython==2.2.1